test(plg): split admin.test.js into 3 files + hoist esmock to before (SITES-46495)
Split the monolithic 3734-line admin.test.js (143 tests, ~95s wall-clock) into
three focused files and eliminated per-test esmock overhead by hoisting the
module graph instrumentation from beforeEach to before:

- admin-get-status.test.js  — getStatus + admin bypass + getAllOnboardings (39 tests)
- admin-update.test.js      — update / bypass flows (61 tests)
- admin-manage.test.js      — transitionStatus + createOnboarding/updateOnboarding/deleteOnboarding (43 tests)
- admin-esmock-factory.js   — shared esmock helper to eliminate copy-paste of the
                              20-module mock map across multiple files
